Tweaked

+++ code/modules/hydroponics/growninedible.dm
@@ -264,21 +264,14 @@
 		var/stun = Clamp(potency / 10, 1, 10)
 		var/weaken = Clamp(potency / 20, 0.5, 5)
 		M.slip(stun, weaken, src)
+		return 1
 
-//===================banana shoes peels================================
-/obj/item/weapon/grown/bananapeel/specialpeel //used by /obj/item/clothing/shoes/clown_shoes/banana_shoes
+/obj/item/weapon/grown/bananapeel/specialpeel     //used by /obj/item/clothing/shoes/clown_shoes/banana_shoes
 	name = "synthesized banana peel"
 	desc = "A synthetic banana peel."
 
-/obj/item/weapon/grown/bananapeel/specialpeel/New(newloc, new_potency = 5)
-	..()
-
 /obj/item/weapon/grown/bananapeel/specialpeel/Crossed(AM)
-	..()
-	if(istype(AM, /mob/living/carbon)) //aahhhh
-		qdel(src)
-
-//=====================================================================
+	if(..())	qdel(src)
 
 /obj/item/weapon/grown/corncob
 	name = "corn cob"
